+++ src/index.js
@@ -6,6 +6,8 @@ import './imgs/refresh.png';
 import Task from './modules/constructor.js';
 import addtasklocalstorage from './modules/localstorage.js';
 import Methods from './modules/Methods.js';
+import deletetask from './modules/deletetask';
+import clearallsuccess from './modules/clearallsucces';
 
 const ui = new Methods();
 
@@ -20,16 +22,6 @@ resetTodo.addEventListener('click', () => {
   window.location.reload();
 });
 
-let deletetask = (e) =>{
-  if (e.target.name === 'deletetask') {
-    const Deleteid = parseInt(e.target.getAttribute('id'));
-    const tasks = JSON.parse(localStorage.getItem('tasks'));
-    const tasks2 = tasks.filter((task) => task.id !== Deleteid);
-    localStorage.setItem('tasks', JSON.stringify(tasks2));
-    const element = document.getElementById(`${Deleteid}`);
-    element.parentElement.parentElement.remove();
-  }
-}
 
 listtask.addEventListener('click', deletetask);
 
@@ -43,7 +35,6 @@ let reloadcontent = (tasks) =>{
       ui.createelement(description);
     });
   }
-
 }
 
 document.addEventListener('DOMContentLoaded', reloadcontent);
@@ -80,11 +71,6 @@ listtask.addEventListener('click', (e) => {
   localStorage.setItem('tasks', JSON.stringify(currentData));
 });
 
-clearallcomplete.addEventListener('click', () => {
-  const tasks7 = JSON.parse(localStorage.getItem('tasks'));
-  const tasks8 = tasks7.filter((task) => task.completed !== true);
-  localStorage.setItem('tasks', JSON.stringify(tasks8));
-  const checkbox = document.querySelector('input[job="complete"]:checked');
-  checkbox.parentElement.parentElement.remove();
-  window.location.reload();
-});
+
+
+clearallcomplete.addEventListener('click', clearallsuccess);
